Add excerpt or content to RSS Category Headlines widget
https://basecamp.com/1756504/projects/6012819/todos/129007040

Added the excerpt with a fall back to the content trimmed to the WordPress default 55 words. The excerpt length option that is being used in the RSS Feed with Images widget is not working here and I am not sure why. Also, I imagine that they will want the ability to decide if they want these fields optional:

- source url
- excerpt